# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LIS",
        "description": "Laboratory Information Service API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Local server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/lis-service/lab-types/{id}": {
            "get": {
                "tags": [
                    "Lab types"
                ],
                "summary": "Get a lab type by id",
                "operationId": "getById_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabTypeResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Lab types"
                ],
                "summary": "Update a lab type",
                "operationId": "update_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLabType"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabTypeResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lab types"
                ],
                "summary": "Delete a lab type",
                "operationId": "delete_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/v1/lis-service/lab-types": {
            "get": {
                "tags": [
                    "Lab types"
                ],
                "summary": "List all lab types",
                "operationId": "list_3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LabTypeResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lab types"
                ],
                "summary": "Create a lab type",
                "operationId": "create_3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLabType"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabTypeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/lab-results/{id}/reject": {
            "post": {
                "tags": [
                    "Lab results"
                ],
                "summary": "Reject an unsolicited result \u2014 not forwarded; notifies the facility",
                "operationId": "reject",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewDecisionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/lab-results/{id}/accept": {
            "post": {
                "tags": [
                    "Lab results"
                ],
                "summary": "Accept an unsolicited result \u2014 forwards it downstream and notifies the facility",
                "operationId": "accept",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/lis-service/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Lab-workflow statistics for a facility over a time period",
                "description": "Counts orders sent, orders acknowledged, and results received (total / valid / unsolicited) for one facility. The period defaults to the last 30 days when `from`/`to` are omitted; both bounds are inclusive dates.",
                "operationId": "facilityStatistics",
                "parameters": [
                    {
                        "name": "mflCode",
                        "in": "query",
                        "description": "Facility MFL code to scope the statistics to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 101010
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start of the window (inclusive, ISO date). Defaults to 29 days before `to`.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End of the window (inclusive, ISO date). Defaults to today (UTC).",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatisticsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/lis-service/lab-results/{id}": {
            "get": {
                "tags": [
                    "Lab results"
                ],
                "summary": "Get a lab result (with observations)",
                "operationId": "getById_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/lab-results/pending-review": {
            "get": {
                "tags": [
                    "Lab results"
                ],
                "summary": "List unsolicited lab results awaiting review at the current user's facilities",
                "operationId": "pendingReview",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseLabResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateLabType": {
                "type": "object",
                "description": "Update a lab type",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "LabTypeResponse": {
                "type": "object",
                "description": "Lab type",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CreateLabType": {
                "type": "object",
                "description": "Create a lab type",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "name"
                ]
            },
            "ReviewDecisionRequest": {
                "type": "object",
                "description": "Clinician review decision (note optional, typically for rejections)",
                "properties": {
                    "note": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 0
                    }
                }
            },
            "LabResultResponse": {
                "type": "object",
                "description": "A received lab result",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "messageControlId": {
                        "type": "string"
                    },
                    "placerOrderNumber": {
                        "type": "string"
                    },
                    "fillerOrderNumber": {
                        "type": "string"
                    },
                    "labCode": {
                        "type": "string"
                    },
                    "orderingMflCode": {
                        "type": "string"
                    },
                    "orderingHmisCode": {
                        "type": "string"
                    },
                    "patientIdentifier": {
                        "type": "string"
                    },
                    "patientName": {
                        "type": "string"
                    },
                    "patientDob": {
                        "type": "string",
                        "format": "date"
                    },
                    "patientSex": {
                        "type": "string"
                    },
                    "testLoinc": {
                        "type": "string"
                    },
                    "testName": {
                        "type": "string"
                    },
                    "resultStatus": {
                        "type": "string"
                    },
                    "messageKind": {
                        "type": "string"
                    },
                    "specimenCollectedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "reconciliationStatus": {
                        "type": "string"
                    },
                    "matchMethod": {
                        "type": "string"
                    },
                    "candidateCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "reviewStatus": {
                        "type": "string"
                    },
                    "reviewedBy": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "reviewedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "reviewNote": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isCurrent": {
                        "type": "boolean"
                    },
                    "forwardStatus": {
                        "type": "string"
                    },
                    "receivedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "observations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ObservationResponse"
                        }
                    }
                }
            },
            "ObservationResponse": {
                "type": "object",
                "description": "A single observation (OBX) of a lab result",
                "properties": {
                    "setId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "valueType": {
                        "type": "string"
                    },
                    "loinc": {
                        "type": "string"
                    },
                    "localCode": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "numericValue": {
                        "type": "number"
                    },
                    "units": {
                        "type": "string"
                    },
                    "referenceRange": {
                        "type": "string"
                    },
                    "abnormalFlags": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "observedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "StatisticsResponse": {
                "type": "object",
                "description": "Basic lab-workflow statistics for a facility over a time period",
                "properties": {
                    "mflCode": {
                        "type": "string",
                        "description": "Facility MFL code the statistics are scoped to",
                        "example": "101010"
                    },
                    "facilityName": {
                        "type": "string",
                        "description": "Facility name, when the MFL code resolves to a known facility"
                    },
                    "from": {
                        "type": "string",
                        "format": "date",
                        "description": "Start of the reporting window (inclusive)",
                        "example": "2026-06-14"
                    },
                    "to": {
                        "type": "string",
                        "format": "date",
                        "description": "End of the reporting window (inclusive)",
                        "example": "2026-07-13"
                    },
                    "labOrdersSent": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lab orders enqueued for publishing by this facility"
                    },
                    "labOrdersAcknowledged": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lab orders that received an acknowledgement from the laboratory"
                    },
                    "labResultsReceived": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total lab result messages received for this facility"
                    },
                    "labResultsValid": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Received results carrying valid observations (message_kind = RESULT)"
                    },
                    "labResultsUnsolicited": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Received results that could not be matched to a known order (unsolicited)"
                    }
                }
            },
...
            "PagedResponseLabResultResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/LabResultResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}